Rename and move packages
