+++ util/grpc/dialer_linux.go
@@ -8,11 +8,11 @@ import (
 
 	"google.golang.org/grpc"
 
-	netpkg "github.com/netbirdio/netbird/pkg/net"
+	nbnet "github.com/netbirdio/netbird/util/net"
 )
 
 func NewCustomDialer() grpc.DialOption {
 	return grpc.WithContextDialer(func(ctx context.Context, addr string) (net.Conn, error) {
-		return netpkg.NewDialer().DialContext(ctx, "tcp", addr)
+		return nbnet.NewDialer().DialContext(ctx, "tcp", addr)
 	})
 }
